refactor(ui): split useAppStore.ts into per-domain slice files
The 2,097-line monolithic Zustand store already partitioned itself into
~32 commented sections; promote each cohesive group into a typed
StateCreator slice under src/ui/src/stores/slices/ and reduce
useAppStore.ts to a thin combiner. Slices are typed against the full
AppState so cross-slice cascades (removeRepository -> terminal/diff
cleanup, rollbackConversation -> chat/agent/checkpoint reset, etc.)
keep working without indirection.

Behavior is unchanged: AppState is the intersection of the slice
interfaces, so the flat store shape consumers and tests rely on
(50+ component selectors, 76 setState calls in vitest specs) is
preserved exactly.

Refs #491.
