+++ src/ui/src/stores/slices/settingsSlice.ts
@@ -1,0 +1,92 @@
+import type { StateCreator } from "zustand";
+import { DEFAULT_THEME_ID, DEFAULT_LIGHT_THEME_ID } from "../../styles/themes";
+import type { AppState } from "../useAppStore";
+
+export interface SettingsSlice {
+  worktreeBaseDir: string;
+  setWorktreeBaseDir: (dir: string) => void;
+  defaultBranches: Record<string, string>;
+  setDefaultBranches: (branches: Record<string, string>) => void;
+  terminalFontSize: number;
+  setTerminalFontSize: (size: number) => void;
+  uiFontSize: number;
+  setUiFontSize: (size: number) => void;
+  fontFamilySans: string;
+  setFontFamilySans: (font: string) => void;
+  fontFamilyMono: string;
+  setFontFamilyMono: (font: string) => void;
+  systemFonts: string[];
+  setSystemFonts: (fonts: string[]) => void;
+  currentThemeId: string;
+  setCurrentThemeId: (id: string) => void;
+  themeMode: "light" | "dark" | "system";
+  setThemeMode: (mode: "light" | "dark" | "system") => void;
+  themeDark: string;
+  setThemeDark: (id: string) => void;
+  themeLight: string;
+  setThemeLight: (id: string) => void;
+
+  // Sidebar display preferences
+  /// Show running terminal commands under each workspace in the sidebar.
+  /// Off by default — opt in via Settings → Appearance.
+  showSidebarRunningCommands: boolean;
+  setShowSidebarRunningCommands: (v: boolean) => void;
+
+  // Experimental
+  usageInsightsEnabled: boolean;
+  setUsageInsightsEnabled: (enabled: boolean) => void;
+  pluginManagementEnabled: boolean;
+  setPluginManagementEnabled: (enabled: boolean) => void;
+  disable1mContext: boolean;
+  setDisable1mContext: (v: boolean) => void;
+}
+
+export const createSettingsSlice: StateCreator<
+  AppState,
+  [],
+  [],
+  SettingsSlice
+> = (set) => ({
+  worktreeBaseDir: "",
+  setWorktreeBaseDir: (dir) => set({ worktreeBaseDir: dir }),
+  defaultBranches: {},
+  setDefaultBranches: (branches) => set({ defaultBranches: branches }),
+  terminalFontSize: 11,
+  setTerminalFontSize: (size) => set({ terminalFontSize: size }),
+  uiFontSize: 13,
+  setUiFontSize: (size) => set({ uiFontSize: size }),
+  fontFamilySans: "",
+  setFontFamilySans: (font) => set({ fontFamilySans: font }),
+  fontFamilyMono: "",
+  setFontFamilyMono: (font) => set({ fontFamilyMono: font }),
+  systemFonts: [],
+  setSystemFonts: (fonts) => set({ systemFonts: fonts }),
+  currentThemeId: DEFAULT_THEME_ID,
+  setCurrentThemeId: (id) => set({ currentThemeId: id }),
+  themeMode: "dark",
+  setThemeMode: (mode) => set({ themeMode: mode }),
+  themeDark: DEFAULT_THEME_ID,
+  setThemeDark: (id) => set({ themeDark: id }),
+  themeLight: DEFAULT_LIGHT_THEME_ID,
+  setThemeLight: (id) => set({ themeLight: id }),
+
+  showSidebarRunningCommands: false,
+  setShowSidebarRunningCommands: (v) => set({ showSidebarRunningCommands: v }),
+
+  usageInsightsEnabled: false,
+  setUsageInsightsEnabled: (enabled) => set({ usageInsightsEnabled: enabled }),
+  pluginManagementEnabled: false,
+  setPluginManagementEnabled: (enabled) =>
+    set((state) => ({
+      pluginManagementEnabled: enabled,
+      settingsSection:
+        !enabled && state.settingsSection === "claude-code-plugins"
+          ? "experimental"
+          : state.settingsSection,
+      pluginSettingsIntent: enabled ? state.pluginSettingsIntent : null,
+      pluginSettingsRepoId: enabled ? state.pluginSettingsRepoId : null,
+      pluginSettingsTab: enabled ? state.pluginSettingsTab : "available",
+    })),
+  disable1mContext: false,
+  setDisable1mContext: (v) => set({ disable1mContext: v }),
+});
