+++ src/ui/src/stores/useAppStore.ts
@@ -1,2132 +1,104 @@
 import { create } from "zustand";
-import { DEFAULT_THEME_ID, DEFAULT_LIGHT_THEME_ID } from "../styles/themes";
-import { debugChat } from "../utils/chatDebug";
-import { extractLatestCallUsage } from "../utils/extractLatestCallUsage";
 import {
-  extractCompactionEvents,
-  type CompactionEvent,
-} from "../utils/compactionSentinel";
-import type {
-  Repository,
-  Workspace,
-  ChatMessage,
-  ChatAttachment,
-  AttachmentInput,
-  ChatSession,
-  DiffFile,
-  DiffFileTab,
-  FileDiff,
-  DiffViewMode,
-  TerminalTab,
-  TerminalPaneNode,
-  TerminalPaneNodeId,
-  TerminalSplitDirection,
-  RemoteConnectionInfo,
-  DiscoveredServer,
-  ConversationCheckpoint,
-} from "../types";
+  createAgentInteractionSlice,
+  type AgentInteractionSlice,
+  type AgentQuestion,
+  type AgentQuestionItem,
+  type ChatSearchState,
+  type PlanApproval,
+} from "./slices/agentInteractionSlice";
 import {
-  allLeafIds as allPaneLeafIds,
-  closeLeaf as closeLeafInTree,
-  countLeaves as countPaneLeaves,
-  makeLeaf as makePaneLeaf,
-  splitLeaf as splitLeafInTree,
-  updateSizes as updateSizesInTree,
-} from "./terminalPaneTree";
-import type { McpStatusSnapshot } from "../types/mcp";
-import type { RemoteInitialData } from "../types/remote";
-import type { DetectedApp } from "../types/apps";
-import type { ClaudeCodeUsage } from "../types/usage";
-import type {
-  AnalyticsMetrics,
-  DashboardMetrics,
-  WorkspaceMetrics,
-} from "../types/metrics";
+  createChatSessionsSlice,
+  type ChatSessionsSlice,
+} from "./slices/chatSessionsSlice";
 import {
-  getAnalyticsMetrics,
-  getDashboardMetrics,
-  getWorkspaceMetricsBatch,
-} from "../services/tauri";
-import type { FileContent, SlashCommand } from "../services/tauri";
-import type {
-  PluginSettingsIntent,
-  PluginSettingsTab,
-} from "../types/plugins";
-
-export type PermissionLevel = "readonly" | "standard" | "full";
-
-/** Return a new sessionsByWorkspace map with `needs_attention` / `attention_kind`
- *  cleared for the given session id. Returns the original map when the session
- *  isn't found so callers can compare referential equality. */
-function clearSessionAttention(
-  map: Record<string, ChatSession[]>,
-  sessionId: string,
-): Record<string, ChatSession[]> {
-  for (const [wsId, sessions] of Object.entries(map)) {
-    const idx = sessions.findIndex((s) => s.id === sessionId);
-    if (idx >= 0) {
-      const updated = [...sessions];
-      updated[idx] = {
-        ...updated[idx],
-        needs_attention: false,
-        attention_kind: null,
-      };
-      return { ...map, [wsId]: updated };
-    }
-  }
-  return map;
-}
-
-export interface ToolActivity {
-  toolUseId: string;
-  toolName: string;
-  inputJson: string;
-  resultText: string;
-  collapsed: boolean;
-  summary: string;
-}
-
-export interface CompletedTurn {
-  id: string;
-  activities: ToolActivity[];
-  messageCount: number;
-  collapsed: boolean;
-  /** Index into chatMessages at the time of finalization — used to render
-   *  the turn summary at the correct chronological position. */
-  afterMessageIndex: number;
-  /** Commit hash from the corresponding conversation checkpoint, if any.
-   *  Used to gate the "fork workspace at this turn" action. */
-  commitHash?: string | null;
-  /** Total time this turn took, in milliseconds. Summed from the
-   *  duration_ms of assistant messages produced during the turn. */
-  durationMs?: number;
-  /** Turn-total input tokens. Live turns receive this from the CLI's
-   *  `result.usage`; persisted turns are reconstructed by summing the
-   *  `input_tokens` of each assistant `ChatMessage` in the turn (see
-   *  `reconstructCompletedTurns`). Undefined for legacy turns with no
-   *  token metadata on any message. */
-  inputTokens?: number;
-  /** Turn-total output tokens. Live turns receive this from the CLI's
-   *  `result.usage`; persisted turns are reconstructed by summing the
-   *  `output_tokens` of each assistant `ChatMessage` in the turn. */
-  outputTokens?: number;
-  /** Turn-total cache-read tokens. Live turns receive this from the CLI's
-   *  `result.usage.cache_read_input_tokens`. Persisted turns use the MAX
-   *  (not sum) of per-message `cache_read_tokens` — cache counts are
-   *  cumulative-per-API-call, so summing across a multi-message tool-use
-   *  turn would double-count the shared prompt prefix each call re-reads. */
-  cacheReadTokens?: number;
-  /** Turn-total cache-creation tokens. Same max-based reconstruction semantics
-   *  as `cacheReadTokens`. */
-  cacheCreationTokens?: number;
-}
-
-export interface AgentQuestionItem {
-  header?: string;
-  question: string;
-  options: Array<{ label: string; description?: string }>;
-  multiSelect?: boolean;
-}
-
-export interface AgentQuestion {
-  sessionId: string;
-  toolUseId: string;
-  questions: AgentQuestionItem[];
-}
-
-export interface PlanApproval {
-  sessionId: string;
-  toolUseId: string;
-  planFilePath: string | null;
-  allowedPrompts: Array<{ tool: string; prompt: string }>;
-}
-
-/**
- * Per-workspace state for the in-chat Cmd/Ctrl+F search bar.
- * `query` is preserved when the bar is closed so re-opening with the same
- * workspace selected restores the previous search.
- * `matchIndex` is the active match's 0-based index across all hits in the
- * current workspace; -1 when there are no matches yet.
- */
-export interface ChatSearchState {
-  open: boolean;
-  query: string;
-  matchIndex: number;
-}
-
-/**
- * Token usage from the most recent completed turn for a workspace.
- * Lives as its own slice (`latestTurnUsage`) rather than being derived from
- * `completedTurns` because `finalizeTurn` early-returns for tool-free turns
- * — so a Q&A turn without tool calls doesn't add a CompletedTurn but should
- * still refresh the ContextMeter. The shape matches the `result.usage` block
- * the CLI emits on every turn end.
- */
-export interface TurnUsage {
-  inputTokens?: number;
-  outputTokens?: number;
-  cacheReadTokens?: number;
-  cacheCreationTokens?: number;
-}
-
-interface AppState {
-  // -- Repositories --
-  repositories: Repository[];
-  setRepositories: (repos: Repository[]) => void;
-  addRepository: (repo: Repository) => void;
-  updateRepository: (id: string, updates: Partial<Repository>) => void;
-  removeRepository: (id: string) => void;
-
-  // -- Workspaces --
-  workspaces: Workspace[];
-  selectedWorkspaceId: string | null;
-  setWorkspaces: (workspaces: Workspace[]) => void;
-  addWorkspace: (ws: Workspace) => void;
-  updateWorkspace: (id: string, updates: Partial<Workspace>) => void;
-  removeWorkspace: (id: string) => void;
-  selectWorkspace: (id: string | null) => void;
-
-  // -- Chat sessions (tabs) --
-  sessionsByWorkspace: Record<string, ChatSession[]>;
-  selectedSessionIdByWorkspaceId: Record<string, string>;
-  setSessionsForWorkspace: (wsId: string, sessions: ChatSession[]) => void;
-  addChatSession: (session: ChatSession) => void;
-  updateChatSession: (
-    sessionId: string,
-    updates: Partial<ChatSession>,
-  ) => void;
-  removeChatSession: (sessionId: string) => void;
-  selectSession: (workspaceId: string, sessionId: string) => void;
-
-  // -- Chat --
-  chatMessages: Record<string, ChatMessage[]>;
-  chatAttachments: Record<string, ChatAttachment[]>;
-  setChatAttachments: (wsId: string, attachments: ChatAttachment[]) => void;
-  addChatAttachments: (wsId: string, attachments: ChatAttachment[]) => void;
-  streamingContent: Record<string, string>;
-  streamingThinking: Record<string, string>;
-  pendingTypewriter: Record<string, { messageId: string; text: string } | null>;
-  showThinkingBlocks: Record<string, boolean>;
-  toolActivities: Record<string, ToolActivity[]>;
-  completedTurns: Record<string, CompletedTurn[]>;
-  /** Latest `result.usage` values per workspace — kept in sync with every
-   *  turn end, including tool-free turns that don't produce a CompletedTurn.
-   *  The ContextMeter reads from here so it reflects the latest turn even
-   *  when the timeline doesn't record one. */
-  latestTurnUsage: Record<string, TurnUsage>;
-  setLatestTurnUsage: (wsId: string, usage: TurnUsage) => void;
-  /** Delete the meter's usage entry for a workspace. Used when a
-   *  rollback or empty load leaves no assistant message with token data —
-   *  clearing hides the meter rather than leaving a stale value. */
-  clearLatestTurnUsage: (wsId: string) => void;
-  promptStartTime: Record<string, number>;
-  setPromptStartTime: (wsId: string, time: number) => void;
-  clearPromptStartTime: (wsId: string) => void;
-  /** Per-workspace compaction history, re-derived from the persisted
-   *  COMPACTION:* sentinel messages on workspace load and updated live
-   *  on compact_boundary events. This slice stores derived metadata
-   *  for future consumers (e.g. a compaction counter); ChatPanel's
-   *  divider rendering dispatches on persisted sentinel System messages
-   *  in `chatMessages[wsId]` rather than reading this slice directly. */
-  compactionEvents: Record<string, CompactionEvent[]>;
-  setCompactionEvents: (wsId: string, events: CompactionEvent[]) => void;
-  addCompactionEvent: (wsId: string, event: CompactionEvent) => void;
-  setChatMessages: (wsId: string, messages: ChatMessage[]) => void;
-  addChatMessage: (wsId: string, message: ChatMessage) => void;
-  setStreamingContent: (wsId: string, content: string) => void;
-  appendStreamingContent: (wsId: string, text: string) => void;
-  setPendingTypewriter: (wsId: string, messageId: string, text: string) => void;
-  /** Atomic drain-end handoff: clears both `pendingTypewriter` and
-   *  `streamingThinking` in a single store update so the streaming thinking
-   *  block and the draining assistant text hand off to the completed message
-   *  in the same render, without a gap or a 1-frame duplicate. */
-  finishTypewriterDrain: (wsId: string) => void;
-  appendStreamingThinking: (wsId: string, text: string) => void;
-  clearStreamingThinking: (wsId: string) => void;
-  setShowThinkingBlocks: (wsId: string, show: boolean) => void;
-  setToolActivities: (wsId: string, activities: ToolActivity[]) => void;
-  addToolActivity: (wsId: string, activity: ToolActivity) => void;
-  updateToolActivity: (
-    wsId: string,
-    toolUseId: string,
-    updates: Partial<ToolActivity>,
-  ) => void;
-  toggleToolActivityCollapsed: (wsId: string, index: number) => void;
-  finalizeTurn: (
-    wsId: string,
-    messageCount: number,
-    turnId?: string,
-    durationMs?: number,
-    inputTokens?: number,
-    outputTokens?: number,
-    cacheReadTokens?: number,
-    cacheCreationTokens?: number,
-  ) => void;
-  hydrateCompletedTurns: (wsId: string, turns: CompletedTurn[]) => void;
-  setCompletedTurns: (wsId: string, turns: CompletedTurn[]) => void;
-  toggleCompletedTurn: (wsId: string, turnIndex: number) => void;
-  appendToolActivityInput: (
-    wsId: string,
-    toolUseId: string,
-    partialJson: string,
-  ) => void;
-
-  // -- Agent Questions (per-session) --
-  agentQuestions: Record<string, AgentQuestion>;
-  setAgentQuestion: (q: AgentQuestion) => void;
-  clearAgentQuestion: (sessionId: string) => void;
-
-  // -- Plan Approvals (per-session) --
-  planApprovals: Record<string, PlanApproval>;
-  setPlanApproval: (p: PlanApproval) => void;
-  clearPlanApproval: (sessionId: string) => void;
-
-  // -- Chat Search (per-workspace) --
-  chatSearch: Record<string, ChatSearchState>;
-  openChatSearch: (wsId: string) => void;
-  closeChatSearch: (wsId: string) => void;
-  setChatSearchQuery: (wsId: string, query: string) => void;
-  setChatSearchMatchIndex: (wsId: string, idx: number) => void;
-
-  // -- Queued Messages (sent while agent is running, dispatched when idle) --
-  queuedMessages: Record<
-    string,
-    {
-      content: string;
-      mentionedFiles?: string[];
-      attachments?: AttachmentInput[];
-    }
-  >;
-  setQueuedMessage: (
-    wsId: string,
-    content: string,
-    mentionedFiles?: string[],
-    attachments?: AttachmentInput[],
-  ) => void;
-  clearQueuedMessage: (wsId: string) => void;
-
-  // -- Checkpoints --
-  checkpoints: Record<string, ConversationCheckpoint[]>;
-  setCheckpoints: (wsId: string, cps: ConversationCheckpoint[]) => void;
-  addCheckpoint: (wsId: string, cp: ConversationCheckpoint) => void;
-  rollbackConversation: (
-    sessionId: string,
-    workspaceId: string,
-    checkpointId: string,
-    messages: ChatMessage[],
-  ) => void;
-
-  // -- Notifications --
-  unreadCompletions: Set<string>; // workspace IDs with unread completions
-  markWorkspaceAsUnread: (wsId: string) => void;
-  clearUnreadCompletion: (wsId: string) => void;
-
-  // -- Toasts --
-  toasts: { id: string; message: string }[];
-  addToast: (message: string) => void;
-  removeToast: (id: string) => void;
-
-  // -- Permissions --
-  permissionLevel: Record<string, PermissionLevel>;
-  setPermissionLevel: (wsId: string, level: PermissionLevel) => void;
-
-  // -- Toolbar --
-  selectedModel: Record<string, string>;
-  fastMode: Record<string, boolean>;
-  thinkingEnabled: Record<string, boolean>;
-  planMode: Record<string, boolean>;
-  effortLevel: Record<string, string>;
-  chromeEnabled: Record<string, boolean>;
-  modelSelectorOpen: boolean;
-  setSelectedModel: (wsId: string, model: string) => void;
-  setFastMode: (wsId: string, enabled: boolean) => void;
-  setThinkingEnabled: (wsId: string, enabled: boolean) => void;
-  setPlanMode: (wsId: string, enabled: boolean) => void;
-  setEffortLevel: (wsId: string, level: string) => void;
-  setChromeEnabled: (wsId: string, enabled: boolean) => void;
-  setModelSelectorOpen: (open: boolean) => void;
-
-  // -- Diff --
-  diffFiles: DiffFile[];
-  diffMergeBase: string | null;
-  diffSelectedFile: string | null;
-  diffSelectedLayer: import("../types/diff").DiffLayer | null;
-  diffStagedFiles: import("../types/diff").StagedDiffFiles | null;
-  diffContent: FileDiff | null;
-  diffViewMode: DiffViewMode;
-  diffLoading: boolean;
-  diffError: string | null;
-  // Markdown-preview overlay for the active diff tab. Resets to "diff" on
-  // every tab switch (no per-tab persistence). Only meaningful when the
-  // selected file's extension is .md/.markdown — UI hides the toggle
-  // otherwise. Content is the working-tree version (post-edit), fetched
-  // separately from the diff so the user sees how their changes will render.
-  diffPreviewMode: "diff" | "rendered";
-  diffPreviewContent: FileContent | null;
-  diffPreviewLoading: boolean;
-  diffPreviewError: string | null;
-  // Per-workspace open diff-file tabs. Ephemeral — not persisted across
-  // restarts. Identity is (path, layer); the same path opened from two
-  // different layers produces two distinct tabs because their diff content
-  // differs.
-  diffTabsByWorkspace: Record<string, DiffFileTab[]>;
-  setDiffFiles: (files: DiffFile[], mergeBase: string, stagedFiles?: import("../types/diff").StagedDiffFiles | null) => void;
-  setDiffSelectedFile: (path: string | null, layer?: import("../types/diff").DiffLayer | null) => void;
-  setDiffContent: (content: FileDiff | null) => void;
-  setDiffViewMode: (mode: DiffViewMode) => void;
-  setDiffLoading: (loading: boolean) => void;
-  setDiffError: (error: string | null) => void;
-  setDiffPreviewMode: (mode: "diff" | "rendered") => void;
-  setDiffPreviewContent: (content: FileContent | null) => void;
-  setDiffPreviewLoading: (loading: boolean) => void;
-  setDiffPreviewError: (error: string | null) => void;
-  clearDiff: () => void;
-  // Open a diff tab for the given file (deduped by path+layer) and make it
-  // the active view. The previously-selected chat session stays selected so
-  // closing all diff tabs restores it.
-  openDiffTab: (workspaceId: string, path: string, layer?: import("../types/diff").DiffLayer | null) => void;
-  // Focus an already-open diff tab without mutating the tab list.
-  selectDiffTab: (path: string, layer?: import("../types/diff").DiffLayer | null) => void;
-  // Close a diff tab. If it was the active diff, the chat session selected
-  // for this workspace becomes active again.
-  closeDiffTab: (workspaceId: string, path: string, layer?: import("../types/diff").DiffLayer | null) => void;
-
-  // -- Terminal --
-  terminalTabs: Record<string, TerminalTab[]>;
-  // Active tab id is workspace-scoped: switching workspaces preserves each
-  // workspace's last-active tab independently.
-  activeTerminalTabId: Record<string, number | null>;
-  terminalPanelVisible: boolean;
-  /// Currently-running foreground commands, keyed by `wsId` then by `ptyId`.
-  /// An entry exists only while that PTY's foreground command is running —
-  /// it's added on `pty-command-detected`, removed on `pty-command-stopped`
-  /// or `pty-exit`. The sidebar renders one row per entry, so a workspace
-  /// with two terminals running `nxv dev` and `sleep 30` shows both.
-  workspaceTerminalCommands: Record<string, Record<number, string | null>>;
-  setTerminalTabs: (wsId: string, tabs: TerminalTab[]) => void;
-  addTerminalTab: (wsId: string, tab: TerminalTab) => void;
-  removeTerminalTab: (wsId: string, tabId: number) => void;
-  setActiveTerminalTab: (wsId: string, id: number | null) => void;
-  toggleTerminalPanel: () => void;
-  setWorkspaceRunningCommand: (
-    wsId: string,
-    ptyId: number,
-    command: string | null,
-  ) => void;
-  clearWorkspaceRunningCommand: (wsId: string, ptyId: number) => void;
-
-  // Per-tab split-pane layout (ephemeral — not persisted). Keyed by tab id.
-  terminalPaneTrees: Record<number, TerminalPaneNode>;
-  // Active leaf id per tab — which pane receives focus and keyboard shortcuts.
-  activeTerminalPaneId: Record<number, TerminalPaneNodeId>;
-  // Maximum leaves per tab before splits are refused. Exposed on state so
-  // the UI can disable the split button at the cap.
-  terminalPaneMaxLeaves: number;
-  // Ensure a tab has a pane tree; no-op if one already exists. Returns the
-  // (possibly pre-existing) root leaf id for convenience.
-  ensurePaneTree: (tabId: number) => TerminalPaneNodeId;
-  setPaneTree: (tabId: number, tree: TerminalPaneNode) => void;
-  // Split the given leaf in two. Returns the id of the newly-created leaf,
-  // or null if the split was refused (cap reached, leaf not in tree, etc).
-  splitPane: (
-    tabId: number,
-    leafId: TerminalPaneNodeId,
-    direction: TerminalSplitDirection,
-  ) => TerminalPaneNodeId | null;
-  // Close a single pane. Returns the id of the newly-focused leaf, or null
-  // if the pane was the sole leaf (caller should close the tab instead).
-  closePane: (
-    tabId: number,
-    leafId: TerminalPaneNodeId,
-  ) => TerminalPaneNodeId | null;
-  setActivePane: (tabId: number, leafId: TerminalPaneNodeId) => void;
-  setPaneSizes: (
-    tabId: number,
-    splitId: TerminalPaneNodeId,
-    sizes: [number, number],
-  ) => void;
-  setPanePtyId: (
-    tabId: number,
-    leafId: TerminalPaneNodeId,
-    ptyId: number,
-  ) => void;
-  setPaneSpawnError: (
-    tabId: number,
-    leafId: TerminalPaneNodeId,
-    error: string | null,
-  ) => void;
-
-  // -- SCM --
-  scmSummary: Record<string, import("../types/plugin").ScmSummary>;
-  scmDetail: import("../types/plugin").ScmDetail | null;
-  scmDetailLoading: boolean;
-  setScmSummary: (wsId: string, summary: import("../types/plugin").ScmSummary) => void;
-  setScmDetail: (detail: import("../types/plugin").ScmDetail | null) => void;
-  setScmDetailLoading: (loading: boolean) => void;
-
-  // -- UI --
-  metaKeyHeld: boolean;
-  setMetaKeyHeld: (held: boolean) => void;
-  sidebarVisible: boolean;
-  rightSidebarVisible: boolean;
-  sidebarWidth: number;
-  rightSidebarWidth: number;
-  terminalHeight: number;
-  rightSidebarTab: "changes" | "tasks" | "scm";
-  sidebarGroupBy: "status" | "repo";
-  sidebarRepoFilter: string; // repo ID or "all"
-  sidebarShowArchived: boolean;
-  repoCollapsed: Record<string, boolean>;
-  statusGroupCollapsed: Record<string, boolean>;
-  fuzzyFinderOpen: boolean;
-  commandPaletteOpen: boolean;
-  toggleSidebar: () => void;
-  toggleRightSidebar: () => void;
-  setRightSidebarTab: (tab: "changes" | "tasks" | "scm") => void;
-  setSidebarWidth: (w: number) => void;
-  setRightSidebarWidth: (w: number) => void;
-  setTerminalHeight: (h: number) => void;
-  setSidebarGroupBy: (g: "status" | "repo") => void;
-  setSidebarRepoFilter: (id: string) => void;
-  setSidebarShowArchived: (show: boolean) => void;
-  toggleRepoCollapsed: (id: string) => void;
-  toggleStatusGroupCollapsed: (id: string) => void;
-  toggleFuzzyFinder: () => void;
-  toggleCommandPalette: () => void;
-
-  // -- Settings page --
-  settingsOpen: boolean;
-  settingsSection: string | null;
-  openSettings: (section?: string) => void;
-  closeSettings: () => void;
-  setSettingsSection: (section: string) => void;
-  pluginSettingsTab: PluginSettingsTab;
-  pluginSettingsRepoId: string | null;
-  pluginSettingsIntent: PluginSettingsIntent | null;
-  pluginRefreshToken: number;
-  openPluginSettings: (intent?: Partial<PluginSettingsIntent>) => void;
-  setPluginSettingsTab: (tab: PluginSettingsTab) => void;
-  setPluginSettingsRepoId: (repoId: string | null) => void;
-  clearPluginSettingsIntent: () => void;
-  bumpPluginRefreshToken: () => void;
-  /** Voice provider whose details panel should be expanded on next render
-   *  of PluginsSettings. Set when the user clicks the mic button and is
-   *  redirected to settings to grant permissions / install models, so the
-   *  user lands directly on the action they need to take. Cleared once the
-   *  panel has consumed the focus. */
-  voiceProviderFocus: string | null;
-  focusVoiceProvider: (providerId: string | null) => void;
-
-  // -- Modals --
-  activeModal: string | null;
-  modalData: Record<string, unknown>;
-  openModal: (name: string, data?: Record<string, unknown>) => void;
-  closeModal: () => void;
-
-  // -- Chat input prefill (e.g. after rollback) --
-  chatInputPrefill: string | null;
-  setChatInputPrefill: (text: string | null) => void;
-  pendingAttachmentsPrefill: AttachmentInput[] | null;
-  setPendingAttachmentsPrefill: (atts: AttachmentInput[] | null) => void;
-
-  // -- Settings --
-  worktreeBaseDir: string;
-  setWorktreeBaseDir: (dir: string) => void;
-  defaultBranches: Record<string, string>;
-  setDefaultBranches: (branches: Record<string, string>) => void;
-  terminalFontSize: number;
-  setTerminalFontSize: (size: number) => void;
-  uiFontSize: number;
-  setUiFontSize: (size: number) => void;
-  fontFamilySans: string;
-  setFontFamilySans: (font: string) => void;
-  fontFamilyMono: string;
-  setFontFamilyMono: (font: string) => void;
-  systemFonts: string[];
-  setSystemFonts: (fonts: string[]) => void;
-  currentThemeId: string;
-  setCurrentThemeId: (id: string) => void;
-  themeMode: "light" | "dark" | "system";
-  setThemeMode: (mode: "light" | "dark" | "system") => void;
-  themeDark: string;
-  setThemeDark: (id: string) => void;
-  themeLight: string;
-  setThemeLight: (id: string) => void;
-  lastMessages: Record<string, ChatMessage>;
-  setLastMessages: (msgs: Record<string, ChatMessage>) => void;
-
-  // -- Remote Connections --
-  remoteConnections: RemoteConnectionInfo[];
-  discoveredServers: DiscoveredServer[];
-  activeRemoteIds: string[];
-  setRemoteConnections: (conns: RemoteConnectionInfo[]) => void;
-  addRemoteConnection: (conn: RemoteConnectionInfo) => void;
-  removeRemoteConnection: (id: string) => void;
-  setDiscoveredServers: (servers: DiscoveredServer[]) => void;
-  setActiveRemoteIds: (ids: string[]) => void;
-  addActiveRemoteId: (id: string) => void;
-  removeActiveRemoteId: (id: string) => void;
-  mergeRemoteData: (connectionId: string, data: RemoteInitialData) => void;
-  clearRemoteData: (connectionId: string) => void;
-
-  // -- Local Server --
-  localServerRunning: boolean;
-  localServerConnectionString: string | null;
-  setLocalServerRunning: (running: boolean) => void;
-  setLocalServerConnectionString: (cs: string | null) => void;
-
-  // -- MCP Status (per-repository) --
-  mcpStatus: Record<string, McpStatusSnapshot>;
-  setMcpStatus: (repoId: string, snapshot: McpStatusSnapshot) => void;
-  clearMcpStatus: (repoId: string) => void;
-
-  // -- Detected Apps --
-  detectedApps: DetectedApp[];
-  setDetectedApps: (apps: DetectedApp[]) => void;
-
-  // -- Sidebar display preferences --
-  /// Show running terminal commands under each workspace in the sidebar.
-  /// Off by default — opt in via Settings → Appearance.
-  showSidebarRunningCommands: boolean;
-  setShowSidebarRunningCommands: (v: boolean) => void;
-
-  // -- Experimental --
-  usageInsightsEnabled: boolean;
-  setUsageInsightsEnabled: (enabled: boolean) => void;
-  pluginManagementEnabled: boolean;
-  setPluginManagementEnabled: (enabled: boolean) => void;
-  disable1mContext: boolean;
-  setDisable1mContext: (v: boolean) => void;
-
-  // -- Claude Code Usage --
-  claudeCodeUsage: ClaudeCodeUsage | null;
-  setClaudeCodeUsage: (usage: ClaudeCodeUsage | null) => void;
-
-  // -- Metrics --
-  dashboardMetrics: DashboardMetrics | null;
-  analyticsMetrics: AnalyticsMetrics | null;
-  workspaceMetrics: Record<string, WorkspaceMetrics>;
-  metricsError: string | null;
-  setDashboardMetrics: (metrics: DashboardMetrics | null) => void;
-  setAnalyticsMetrics: (metrics: AnalyticsMetrics | null) => void;
-  setWorkspaceMetrics: (metrics: Record<string, WorkspaceMetrics>) => void;
-  fetchDashboardMetrics: () => Promise<void>;
-  fetchAnalyticsMetrics: () => Promise<void>;
-  fetchWorkspaceMetricsBatch: (ids: string[]) => Promise<void>;
-
-  // -- Updater --
-  updateAvailable: boolean;
-  updateVersion: string | null;
-  updateDismissed: boolean;
-  updateInstallWhenIdle: boolean;
-  updateDownloading: boolean;
-  updateProgress: number;
-  updateChannel: "stable" | "nightly";
-  updateError: string | null;
-  setUpdateAvailable: (available: boolean, version: string | null) => void;
-  setUpdateDismissed: (dismissed: boolean) => void;
-  setUpdateInstallWhenIdle: (enabled: boolean) => void;
-  setUpdateDownloading: (downloading: boolean) => void;
-  setUpdateProgress: (progress: number) => void;
-  setUpdateChannel: (channel: "stable" | "nightly") => void;
-  setUpdateError: (error: string | null) => void;
-
-  // -- App info --
-  appVersion: string | null;
-  setAppVersion: (version: string | null) => void;
-
-  // -- Slash commands (shared so native dispatch can honor file-based shadows) --
-  slashCommandsByWorkspace: Record<string, SlashCommand[]>;
-  setSlashCommands: (wsId: string, cmds: SlashCommand[]) => void;
-}
-
-const toastTimers = new Map<string, number>();
-
-export const useAppStore = create<AppState>((set, get) => ({
-  // -- Repositories --
-  repositories: [],
-  setRepositories: (repos) => set({ repositories: repos }),
-  addRepository: (repo) =>
-    set((s) => ({ repositories: [...s.repositories, repo] })),
-  updateRepository: (id, updates) =>
-    set((s) => ({
-      repositories: s.repositories.map((r) =>
-        r.id === id ? { ...r, ...updates } : r,
-      ),
-    })),
-  removeRepository: (id) =>
-    set((s) => {
-      const removedWsIds = s.workspaces
-        .filter((w) => w.repository_id === id)
-        .map((w) => w.id);
-      const removedWsIdSet = new Set(removedWsIds);
-      const newTerminalTabs = { ...s.terminalTabs };
-      const newActiveTerminalTabId = { ...s.activeTerminalTabId };
-      const newWorkspaceTerminalCommands = { ...s.workspaceTerminalCommands };
-      const newUnreadCompletions = new Set(s.unreadCompletions);
-      // Collect all tab ids we're about to orphan, then drop their pane
-      // trees and active-pane entries alongside the workspace-keyed maps.
-      const orphanedTabIds = new Set<number>();
-      for (const wsId of removedWsIds) {
-        for (const tab of s.terminalTabs[wsId] ?? []) orphanedTabIds.add(tab.id);
-        delete newTerminalTabs[wsId];
-        delete newActiveTerminalTabId[wsId];
-        delete newWorkspaceTerminalCommands[wsId];
-        newUnreadCompletions.delete(wsId);
-      }
-      const newPaneTrees = { ...s.terminalPaneTrees };
-      const newActivePane = { ...s.activeTerminalPaneId };
-      for (const tabId of orphanedTabIds) {
-        delete newPaneTrees[tabId];
-        delete newActivePane[tabId];
-      }
-      const newDiffTabs = { ...s.diffTabsByWorkspace };
-      for (const wsId of removedWsIds) {
-        delete newDiffTabs[wsId];
-      }
-      return {
-        repositories: s.repositories.filter((r) => r.id !== id),
-        workspaces: s.workspaces.filter((w) => w.repository_id !== id),
-        // If the selected workspace belonged to the removed repo, deselect
-        // it so the rest of the app doesn't point at a vanished id.
-        selectedWorkspaceId:
-          s.selectedWorkspaceId && removedWsIdSet.has(s.selectedWorkspaceId)
-            ? null
-            : s.selectedWorkspaceId,
-        terminalTabs: newTerminalTabs,
-        activeTerminalTabId: newActiveTerminalTabId,
-        workspaceTerminalCommands: newWorkspaceTerminalCommands,
-        unreadCompletions: newUnreadCompletions,
-        terminalPaneTrees: newPaneTrees,
-        activeTerminalPaneId: newActivePane,
-        diffTabsByWorkspace: newDiffTabs,
-      };
-    }),
-
-  // -- Workspaces --
-  workspaces: [],
-  selectedWorkspaceId: null,
-  setWorkspaces: (workspaces) => set({ workspaces }),
-  addWorkspace: (ws) => set((s) => ({ workspaces: [...s.workspaces, ws] })),
-  updateWorkspace: (id, updates) =>
-    set((s) => ({
-      workspaces: s.workspaces.map((w) =>
-        w.id === id ? { ...w, ...updates } : w,
-      ),
-    })),
-  removeWorkspace: (id) =>
-    set((s) => {
-      const newUnreadCompletions = new Set(s.unreadCompletions);
-      newUnreadCompletions.delete(id);
-      // Drop all per-workspace terminal state for the removed workspace.
-      // The cleanup effect in TerminalPanel watches `terminalTabs` and tears
-      // down xterm instances and PTYs whose tab ids no longer exist in any
-      // workspace; the other maps are value-keyed by workspace id.
-      const orphanedTabIds = (s.terminalTabs[id] ?? []).map((t) => t.id);
-      const newTerminalTabs = { ...s.terminalTabs };
-      delete newTerminalTabs[id];
-      const newActiveTerminalTabId = { ...s.activeTerminalTabId };
-      delete newActiveTerminalTabId[id];
-      const newWorkspaceTerminalCommands = { ...s.workspaceTerminalCommands };
-      delete newWorkspaceTerminalCommands[id];
-      const newPaneTrees = { ...s.terminalPaneTrees };
-      const newActivePane = { ...s.activeTerminalPaneId };
-      for (const tabId of orphanedTabIds) {
-        delete newPaneTrees[tabId];
-        delete newActivePane[tabId];
-      }
-      const newDiffTabs = { ...s.diffTabsByWorkspace };
-      delete newDiffTabs[id];
-      return {
-        workspaces: s.workspaces.filter((w) => w.id !== id),
-        selectedWorkspaceId:
-          s.selectedWorkspaceId === id ? null : s.selectedWorkspaceId,
-        unreadCompletions: newUnreadCompletions,
-        terminalTabs: newTerminalTabs,
-        activeTerminalTabId: newActiveTerminalTabId,
-        workspaceTerminalCommands: newWorkspaceTerminalCommands,
-        terminalPaneTrees: newPaneTrees,
-        activeTerminalPaneId: newActivePane,
-        diffTabsByWorkspace: newDiffTabs,
-      };
-    }),
-  selectWorkspace: (id) =>
-    set((s) => {
-      const updates: Partial<AppState> = {
-        selectedWorkspaceId: id,
-        rightSidebarTab: "changes",
-        // diffSelectedFile/Layer are workspace-global pointers; switching
-        // workspaces must drop them so the new workspace's tab strip and
-        // content render cleanly. Per-workspace diff *tabs* live in
-        // diffTabsByWorkspace and are preserved.
-        diffSelectedFile: null,
-        diffSelectedLayer: null,
-        diffContent: null,
-        diffError: null,
-        diffPreviewMode: "diff",
-        diffPreviewContent: null,
-        diffPreviewLoading: false,
-        diffPreviewError: null,
-      };
-      if (id && s.unreadCompletions.has(id)) {
-        const next = new Set(s.unreadCompletions);
-        next.delete(id);
-        updates.unreadCompletions = next;
-      }
-      return updates;
-    }),
-
-  // -- Chat sessions (tabs) --
-  sessionsByWorkspace: {},
-  selectedSessionIdByWorkspaceId: {},
-  setSessionsForWorkspace: (wsId, sessions) =>
-    set((s) => {
-      const next = { ...s.sessionsByWorkspace, [wsId]: sessions };
-      const nextSelected = { ...s.selectedSessionIdByWorkspaceId };
-      const selected = nextSelected[wsId];
-      const activeSessions = sessions.filter((x) => x.status === "Active");
-      const selectedIsActive =
-        selected && activeSessions.some((x) => x.id === selected);
-      if (!selectedIsActive && activeSessions.length > 0) {
-        nextSelected[wsId] = activeSessions[0].id;
-      }
-      return {
-        sessionsByWorkspace: next,
-        selectedSessionIdByWorkspaceId: nextSelected,
-      };
-    }),
-  addChatSession: (session) =>
-    set((s) => {
-      const existing = s.sessionsByWorkspace[session.workspace_id] ?? [];
-      if (existing.some((x) => x.id === session.id)) {
-        return s;
-      }
-      return {
-        sessionsByWorkspace: {
-          ...s.sessionsByWorkspace,
-          [session.workspace_id]: [...existing, session],
-        },
-      };
-    }),
-  updateChatSession: (sessionId, updates) =>
-    set((s) => {
-      for (const [wsId, sessions] of Object.entries(s.sessionsByWorkspace)) {
-        const idx = sessions.findIndex((x) => x.id === sessionId);
-        if (idx >= 0) {
-          const updated = [...sessions];
-          updated[idx] = { ...updated[idx], ...updates };
-          return {
-            sessionsByWorkspace: {
-              ...s.sessionsByWorkspace,
-              [wsId]: updated,
-            },
-          };
-        }
-      }
-      return s;
-    }),
-  removeChatSession: (sessionId) =>
-    set((s) => {
-      const next = { ...s.sessionsByWorkspace };
-      const nextSelected = { ...s.selectedSessionIdByWorkspaceId };
-      for (const [wsId, sessions] of Object.entries(next)) {
-        if (sessions.some((x) => x.id === sessionId)) {
-          next[wsId] = sessions.filter((x) => x.id !== sessionId);
-          if (nextSelected[wsId] === sessionId) {
-            const firstActive = next[wsId].find(
-              (x) => x.status === "Active",
-            );
-            if (firstActive) {
-              nextSelected[wsId] = firstActive.id;
-            } else {
-              delete nextSelected[wsId];
-            }
-          }
-          break;
-        }
-      }
-      return {
-        sessionsByWorkspace: next,
-        selectedSessionIdByWorkspaceId: nextSelected,
-      };
-    }),
-  selectSession: (workspaceId, sessionId) =>
-    set((s) => ({
-      selectedSessionIdByWorkspaceId: {
-        ...s.selectedSessionIdByWorkspaceId,
-        [workspaceId]: sessionId,
-      },
-      // Clicking a chat tab is an explicit "show me chat" intent, so any
-      // active diff view yields. Diff tabs themselves remain in the strip.
-      diffSelectedFile: null,
-      diffSelectedLayer: null,
-      diffPreviewMode: "diff",
-      diffPreviewContent: null,
-      diffPreviewLoading: false,
-      diffPreviewError: null,
-    })),
-
-  // -- Chat --
-  chatMessages: {},
-  chatAttachments: {},
-  setChatAttachments: (wsId, attachments) =>
-    set((s) => ({
-      chatAttachments: { ...s.chatAttachments, [wsId]: attachments },
-    })),
-  addChatAttachments: (wsId, attachments) =>
-    set((s) => ({
-      chatAttachments: {
-        ...s.chatAttachments,
-        [wsId]: [...(s.chatAttachments[wsId] ?? []), ...attachments],
-      },
-    })),
-  streamingContent: {},
-  streamingThinking: {},
-  pendingTypewriter: {},
-  showThinkingBlocks: {},
-  toolActivities: {},
-  completedTurns: {},
-  latestTurnUsage: {},
-  setLatestTurnUsage: (wsId, usage) =>
-    set((s) => ({
-      latestTurnUsage: { ...s.latestTurnUsage, [wsId]: usage },
-    })),
-  clearLatestTurnUsage: (wsId) =>
-    set((s) => {
-      if (!(wsId in s.latestTurnUsage)) return {};
-      const next = { ...s.latestTurnUsage };
-      delete next[wsId];
-      return { latestTurnUsage: next };
-    }),
-  promptStartTime: {},
-  setPromptStartTime: (wsId, time) =>
-    set((s) => ({
-      promptStartTime: { ...s.promptStartTime, [wsId]: time },
-    })),
-  clearPromptStartTime: (wsId) =>
-    set((s) => {
-      if (!(wsId in s.promptStartTime)) return {};
-      const next = { ...s.promptStartTime };
-      delete next[wsId];
-      return { promptStartTime: next };
-    }),
-  compactionEvents: {},
-  setCompactionEvents: (wsId, events) =>
-    set((s) => ({
-      compactionEvents: { ...s.compactionEvents, [wsId]: events },
-    })),
-  addCompactionEvent: (wsId, event) =>
-    set((s) => ({
-      compactionEvents: {
-        ...s.compactionEvents,
-        [wsId]: [...(s.compactionEvents[wsId] ?? []), event],
-      },
-    })),
-  setChatMessages: (wsId, messages) =>
-    set((s) => ({
-      chatMessages: { ...s.chatMessages, [wsId]: messages },
-    })),
-  addChatMessage: (wsId, message) =>
-    set((s) => ({
-      chatMessages: {
-        ...s.chatMessages,
-        [wsId]: [...(s.chatMessages[wsId] || []), message],
-      },
-      lastMessages: { ...s.lastMessages, [wsId]: message },
-    })),
-  setStreamingContent: (wsId, content) =>
-    set((s) => ({
-      streamingContent: { ...s.streamingContent, [wsId]: content },
-    })),
-  appendStreamingContent: (wsId, text) =>
-    set((s) => ({
-      streamingContent: {
-        ...s.streamingContent,
-        [wsId]: (s.streamingContent[wsId] || "") + text,
-      },
-    })),
-  setPendingTypewriter: (wsId, messageId, text) =>
-    set((s) => ({
-      pendingTypewriter: {
-        ...s.pendingTypewriter,
-        [wsId]: { messageId, text },
-      },
-    })),
-  finishTypewriterDrain: (wsId) =>
-    set((s) => ({
-      pendingTypewriter: { ...s.pendingTypewriter, [wsId]: null },
-      streamingThinking: { ...s.streamingThinking, [wsId]: "" },
-    })),
-  appendStreamingThinking: (wsId, text) =>
-    set((s) => ({
-      streamingThinking: {
-        ...s.streamingThinking,
-        [wsId]: (s.streamingThinking[wsId] || "") + text,
-      },
-    })),
-  clearStreamingThinking: (wsId) =>
-    set((s) => ({
-      streamingThinking: { ...s.streamingThinking, [wsId]: "" },
-    })),
-  setShowThinkingBlocks: (wsId, show) =>
-    set((s) => ({
-      showThinkingBlocks: { ...s.showThinkingBlocks, [wsId]: show },
-    })),
-  setToolActivities: (wsId, activities) =>
-    set((s) => ({
-      toolActivities: { ...s.toolActivities, [wsId]: activities },
-    })),
-  addToolActivity: (wsId, activity) =>
-    set((s) => ({
-      toolActivities: {
-        ...s.toolActivities,
-        [wsId]: [...(s.toolActivities[wsId] || []), activity],
-      },
-    })),
-  updateToolActivity: (wsId, toolUseId, updates) =>
-    set((s) => ({
-      toolActivities: {
-        ...s.toolActivities,
-        [wsId]: (s.toolActivities[wsId] || []).map((a) =>
-          a.toolUseId === toolUseId ? { ...a, ...updates } : a,
-        ),
-      },
-    })),
-  toggleToolActivityCollapsed: (wsId, index) =>
-    set((s) => ({
-      toolActivities: {
-        ...s.toolActivities,
-        [wsId]: (s.toolActivities[wsId] || []).map((a, i) =>
-          i === index ? { ...a, collapsed: !a.collapsed } : a,
-        ),
-      },
-    })),
-  appendToolActivityInput: (wsId, toolUseId, partialJson) =>
-    set((s) => ({
-      toolActivities: {
-        ...s.toolActivities,
-        [wsId]: (s.toolActivities[wsId] || []).map((a) =>
-          a.toolUseId === toolUseId
-            ? { ...a, inputJson: a.inputJson + partialJson }
-            : a,
-        ),
-      },
-    })),
-  finalizeTurn: (
-    wsId,
-    messageCount,
-    turnId,
-    durationMs,
-    inputTokens,
-    outputTokens,
-    cacheReadTokens,
-    cacheCreationTokens,
-  ) =>
-    set((s) => {
-      // Phase 2.5: finalizeTurn no longer writes latestTurnUsage. The
-      // meter needs per-call values, not the turn-aggregate we receive
-      // here — useAgentStream's result handler calls setLatestTurnUsage
-      // separately with the correct per-call data. The tokens we DO
-      // receive here stay as CompletedTurn aggregate fields for the
-      // TurnFooter's "turn-total work" view.
-      const activities = s.toolActivities[wsId] || [];
-      if (activities.length === 0) {
-        debugChat("store", "finalizeTurn skipped", {
-          wsId,
-          messageCount,
-          turnId: turnId ?? null,
-          existingCompletedTurnIds: (s.completedTurns[wsId] || []).map(
-            (turn) => turn.id,
-          ),
-        });
-        return {};
-      }
-      const turn: CompletedTurn = {
-        id: turnId ?? crypto.randomUUID(),
-        activities: activities.map((a) => ({
-          toolUseId: a.toolUseId,
-          toolName: a.toolName,
-          inputJson: a.inputJson,
-          resultText: a.resultText,
-          collapsed: true,
-          summary: a.summary,
-        })),
-        messageCount,
-        collapsed: true,
-        afterMessageIndex: (s.chatMessages[wsId] || []).length,
-        durationMs,
-        inputTokens,
-        outputTokens,
-        cacheReadTokens,
-        cacheCreationTokens,
-      };
-      debugChat("store", "finalizeTurn", {
-        wsId,
-        turnId: turn.id,
-        messageCount,
-        afterMessageIndex: turn.afterMessageIndex,
-        toolCount: turn.activities.length,
-        toolUseIds: turn.activities.map((activity) => activity.toolUseId),
-        existingCompletedTurnIds: (s.completedTurns[wsId] || []).map(
-          (existingTurn) => existingTurn.id,
-        ),
-      });
-      return {
-        completedTurns: {
-          ...s.completedTurns,
-          [wsId]: [...(s.completedTurns[wsId] || []), turn],
-        },
-        toolActivities: { ...s.toolActivities, [wsId]: [] },
-      };
-    }),
-  hydrateCompletedTurns: (wsId, turns) =>
-    set((s) => {
-      const existing = s.completedTurns[wsId] || [];
-      const existingById = new Map(existing.map((turn) => [turn.id, turn]));
-      const incomingIds = new Set(turns.map((turn) => turn.id));
-
-      const merged = turns.map((turn) => {
-        const existingTurn = existingById.get(turn.id);
-        if (!existingTurn) return turn;
-
-        const existingActivitiesById = new Map(
-          existingTurn.activities.map((activity) => [
-            activity.toolUseId,
-            activity,
-          ]),
-        );
-
-        return {
-          ...turn,
-          collapsed: existingTurn.collapsed,
-          activities: turn.activities.map((activity) => ({
-            ...activity,
-            collapsed:
-              existingActivitiesById.get(activity.toolUseId)?.collapsed ??
-              activity.collapsed,
-          })),
-        };
-      });
-
-      const pendingTurns = existing.filter((turn) => !incomingIds.has(turn.id));
-      const nextTurns = [...merged, ...pendingTurns].sort(
-        (a, b) => a.afterMessageIndex - b.afterMessageIndex,
-      );
-
-      debugChat("store", "hydrateCompletedTurns", {
-        wsId,
-        existingIds: existing.map((turn) => turn.id),
-        incomingIds: turns.map((turn) => turn.id),
-        pendingIds: pendingTurns.map((turn) => turn.id),
-        nextIds: nextTurns.map((turn) => turn.id),
-      });
-
-      return {
-        completedTurns: {
-          ...s.completedTurns,
-          [wsId]: nextTurns,
-        },
-      };
-    }),
-  setCompletedTurns: (wsId, turns) =>
-    set((s) => {
-      debugChat("store", "setCompletedTurns", {
-        wsId,
-        turnIds: turns.map((turn) => turn.id),
-        previousIds: (s.completedTurns[wsId] || []).map((turn) => turn.id),
-      });
-      return {
-        completedTurns: { ...s.completedTurns, [wsId]: turns },
-      };
-    }),
-  toggleCompletedTurn: (wsId, turnIndex) =>
-    set((s) => ({
-      completedTurns: {
-        ...s.completedTurns,
-        [wsId]: (s.completedTurns[wsId] || []).map((t, i) =>
-          i === turnIndex ? { ...t, collapsed: !t.collapsed } : t,
-        ),
-      },
-    })),
-
-  // -- Agent Questions (per-session) --
-  agentQuestions: {},
-  setAgentQuestion: (q) =>
-    set((s) => ({
-      agentQuestions: { ...s.agentQuestions, [q.sessionId]: q },
-    })),
-  clearAgentQuestion: (sessionId) =>
-    set((s) => {
-      const { [sessionId]: _, ...rest } = s.agentQuestions;
-      // Also clear the corresponding ChatSession attention flag so the tab
-      // icon + sidebar aggregate update immediately, without waiting for a
-      // list_chat_sessions refresh.
-      const nextSessions = clearSessionAttention(s.sessionsByWorkspace, sessionId);
-      return { agentQuestions: rest, sessionsByWorkspace: nextSessions };
-    }),
-
-  // -- Plan Approvals (per-session) --
-  planApprovals: {},
-  setPlanApproval: (p) =>
-    set((s) => ({
-      planApprovals: { ...s.planApprovals, [p.sessionId]: p },
-    })),
-  clearPlanApproval: (sessionId) =>
-    set((s) => {
-      const { [sessionId]: _, ...rest } = s.planApprovals;
-      const nextSessions = clearSessionAttention(s.sessionsByWorkspace, sessionId);
-      return { planApprovals: rest, sessionsByWorkspace: nextSessions };
-    }),
-
-  // -- Chat Search (per-workspace) --
-  chatSearch: {},
-  openChatSearch: (wsId) =>
-    set((s) => {
-      const prev = s.chatSearch[wsId];
-      return {
-        chatSearch: {
-          ...s.chatSearch,
-          [wsId]: {
-            open: true,
-            query: prev?.query ?? "",
-            matchIndex: prev?.matchIndex ?? -1,
-          },
-        },
-      };
-    }),
-  closeChatSearch: (wsId) =>
-    set((s) => {
-      const prev = s.chatSearch[wsId];
-      // Preserve query/matchIndex so re-opening restores the previous search.
-      // Return the existing state reference for no-op paths so Zustand's
-      // identity check skips the listener notification entirely.
-      if (!prev) return s;
-      if (!prev.open) return s;
-      return {
-        chatSearch: {
-          ...s.chatSearch,
-          [wsId]: { ...prev, open: false },
-        },
-      };
-    }),
-  setChatSearchQuery: (wsId, query) =>
-    set((s) => {
-      const prev = s.chatSearch[wsId] ?? {
-        open: true,
-        query: "",
-        matchIndex: -1,
-      };
-      // Reset matchIndex on every query change — the new query produces a
-      // fresh match set, so any prior index is meaningless.
-      return {
-        chatSearch: {
-          ...s.chatSearch,
-          [wsId]: { ...prev, query, matchIndex: -1 },
-        },
-      };
-    }),
-  setChatSearchMatchIndex: (wsId, idx) =>
-    set((s) => {
-      const prev = s.chatSearch[wsId];
-      if (!prev) return s;
-      if (prev.matchIndex === idx) return s;
-      return {
-        chatSearch: {
-          ...s.chatSearch,
-          [wsId]: { ...prev, matchIndex: idx },
-        },
-      };
-    }),
-
-  // -- Queued Messages --
-  queuedMessages: {},
-  setQueuedMessage: (wsId, content, mentionedFiles, attachments) =>
-    set((s) => ({
-      queuedMessages: {
-        ...s.queuedMessages,
-        [wsId]: { content, mentionedFiles, attachments },
-      },
-    })),
-  clearQueuedMessage: (wsId) =>
-    set((s) => {
-      const { [wsId]: _, ...rest } = s.queuedMessages;
-      return { queuedMessages: rest };
-    }),
-
-  // -- Checkpoints --
-  checkpoints: {},
-  setCheckpoints: (wsId, cps) =>
-    set((s) => ({
-      checkpoints: { ...s.checkpoints, [wsId]: cps },
-    })),
-  addCheckpoint: (wsId, cp) =>
-    set((s) => ({
-      checkpoints: {
-        ...s.checkpoints,
-        [wsId]: [...(s.checkpoints[wsId] || []), cp],
-      },
-    })),
-  rollbackConversation: (sessionId, workspaceId, checkpointId, messages) =>
-    set((s) => {
-      const { [sessionId]: _q, ...restQuestions } = s.agentQuestions;
-      const { [sessionId]: _p, ...restApprovals } = s.planApprovals;
-      const { [workspaceId]: _cs, ...restChatSearch } = s.chatSearch;
-      // Update lastMessages so workspace preview cards stay in sync.
-      const lastMsg =
-        messages.length > 0 ? messages[messages.length - 1] : undefined;
-      const { [workspaceId]: _lm, ...restLastMessages } = s.lastMessages;
-      const updatedLastMessages = lastMsg
-        ? { ...s.lastMessages, [workspaceId]: lastMsg }
-        : restLastMessages;
-      // Recompute the meter's latestTurnUsage from the rolled-back message
-      // list. Write if the last assistant message has token data; delete
-      // the entry otherwise so the meter hides.
-      const nextCall = extractLatestCallUsage(messages);
-      let latestTurnUsage = s.latestTurnUsage;
-      if (nextCall) {
-        latestTurnUsage = { ...s.latestTurnUsage, [sessionId]: nextCall };
-      } else if (sessionId in s.latestTurnUsage) {
-        const next = { ...s.latestTurnUsage };
-        delete next[sessionId];
-        latestTurnUsage = next;
-      }
-      const nextCompactionEvents = {
-        ...s.compactionEvents,
-        [sessionId]: extractCompactionEvents(messages),
-      };
-      return {
-        chatMessages: { ...s.chatMessages, [sessionId]: messages },
-        lastMessages: updatedLastMessages,
-        completedTurns: { ...s.completedTurns, [sessionId]: [] },
-        toolActivities: { ...s.toolActivities, [sessionId]: [] },
-        streamingContent: { ...s.streamingContent, [sessionId]: "" },
-        streamingThinking: { ...s.streamingThinking, [sessionId]: "" },
-        agentQuestions: restQuestions,
-        planApprovals: restApprovals,
-        chatSearch: restChatSearch,
-        checkpoints: {
-          ...s.checkpoints,
-          [sessionId]: (() => {
-            const current = s.checkpoints[sessionId] || [];
-            const target = current.find((c) => c.id === checkpointId);
-            // If target not found (e.g. clear-all sentinel), clear everything.
-            if (!target) return [];
-            return current.filter((cp) => cp.turn_index <= target.turn_index);
-          })(),
-        },
-        latestTurnUsage,
-        compactionEvents: nextCompactionEvents,
-      };
-    }),
-
-  // -- Notifications --
-  unreadCompletions: new Set<string>(),
-  markWorkspaceAsUnread: (wsId) =>
-    set((s) => {
-      const newSet = new Set(s.unreadCompletions);
-      newSet.add(wsId);
-      return { unreadCompletions: newSet };
-    }),
-  clearUnreadCompletion: (wsId) =>
-    set((s) => {
-      const newSet = new Set(s.unreadCompletions);
-      newSet.delete(wsId);
-      return { unreadCompletions: newSet };
-    }),
-
-  // -- Toasts --
-  toasts: [],
-  addToast: (message) => {
-    const id = crypto.randomUUID();
-    set((s) => ({ toasts: [...s.toasts, { id, message }] }));
-    const handle = window.setTimeout(() => {
-      toastTimers.delete(id);
-      set((s) => ({ toasts: s.toasts.filter((t) => t.id !== id) }));
-    }, 5000);
-    toastTimers.set(id, handle);
-  },
-  removeToast: (id) => {
-    const handle = toastTimers.get(id);
-    if (handle != null) {
-      window.clearTimeout(handle);
-      toastTimers.delete(id);
-    }
-    set((s) => ({ toasts: s.toasts.filter((t) => t.id !== id) }));
-  },
-
-  // -- Permissions --
-  permissionLevel: {},
-  setPermissionLevel: (wsId, level) =>
-    set((s) => ({
-      permissionLevel: { ...s.permissionLevel, [wsId]: level },
-    })),
-
-  // -- Toolbar --
-  selectedModel: {},
-  fastMode: {},
-  thinkingEnabled: {},
-  planMode: {},
-  effortLevel: {},
-  chromeEnabled: {},
-  modelSelectorOpen: false,
-  setSelectedModel: (wsId, model) =>
-    set((s) => ({
-      selectedModel: { ...s.selectedModel, [wsId]: model },
-    })),
-  setFastMode: (wsId, enabled) =>
-    set((s) => ({
-      fastMode: { ...s.fastMode, [wsId]: enabled },
-    })),
-  setThinkingEnabled: (wsId, enabled) =>
-    set((s) => ({
-      thinkingEnabled: { ...s.thinkingEnabled, [wsId]: enabled },
-    })),
-  setPlanMode: (wsId, enabled) =>
-    set((s) => ({
-      planMode: { ...s.planMode, [wsId]: enabled },
-    })),
-  setEffortLevel: (wsId, level) =>
-    set((s) => ({
-      effortLevel: { ...s.effortLevel, [wsId]: level },
-    })),
-  setChromeEnabled: (wsId, enabled) =>
-    set((s) => ({
-      chromeEnabled: { ...s.chromeEnabled, [wsId]: enabled },
-    })),
-  setModelSelectorOpen: (open) => set({ modelSelectorOpen: open }),
-
-  // -- Diff --
-  diffFiles: [],
-  diffMergeBase: null,
-  diffSelectedFile: null,
-  diffSelectedLayer: null,
-  diffStagedFiles: null,
-  diffContent: null,
-  diffViewMode: "Unified",
-  diffLoading: false,
-  diffError: null,
-  diffPreviewMode: "diff",
-  diffPreviewContent: null,
-  diffPreviewLoading: false,
-  diffPreviewError: null,
-  diffTabsByWorkspace: {},
-  setDiffFiles: (files, mergeBase, stagedFiles) =>
-    set({ diffFiles: files, diffMergeBase: mergeBase, diffStagedFiles: stagedFiles ?? null }),
-  setDiffSelectedFile: (path, layer) => set({ diffSelectedFile: path, diffSelectedLayer: layer ?? null }),
-  setDiffContent: (content) => set({ diffContent: content }),
-  setDiffViewMode: (mode) => set({ diffViewMode: mode }),
-  setDiffLoading: (loading) => set({ diffLoading: loading }),
-  setDiffError: (error) => set({ diffError: error }),
-  setDiffPreviewMode: (mode) => set({ diffPreviewMode: mode }),
-  setDiffPreviewContent: (content) => set({ diffPreviewContent: content }),
-  setDiffPreviewLoading: (loading) => set({ diffPreviewLoading: loading }),
-  setDiffPreviewError: (error) => set({ diffPreviewError: error }),
-  clearDiff: () =>
-    set({
-      diffFiles: [],
-      diffMergeBase: null,
-      diffSelectedFile: null,
-      diffSelectedLayer: null,
-      diffStagedFiles: null,
-      diffContent: null,
-      diffError: null,
-      diffPreviewMode: "diff",
-      diffPreviewContent: null,
-      diffPreviewLoading: false,
-      diffPreviewError: null,
-      diffTabsByWorkspace: {},
-    }),
-  openDiffTab: (workspaceId, path, layer) =>
-    set((s) => {
-      const normalizedLayer = layer ?? null;
-      const existing = s.diffTabsByWorkspace[workspaceId] ?? [];
-      const alreadyOpen = existing.some(
-        (t) => t.path === path && t.layer === normalizedLayer,
-      );
-      const nextTabs = alreadyOpen
-        ? s.diffTabsByWorkspace
-        : {
-            ...s.diffTabsByWorkspace,
-            [workspaceId]: [...existing, { path, layer: normalizedLayer }],
-          };
-      // Only clear content when the selection actually changes — clicking the
-      // already-active tab must not blank the viewer (the loader effect
-      // wouldn't refire on identical deps, leaving the user staring at empty).
-      const isSameSelection =
-        s.diffSelectedFile === path && s.diffSelectedLayer === normalizedLayer;
-      return {
-        diffTabsByWorkspace: nextTabs,
-        diffSelectedFile: path,
-        diffSelectedLayer: normalizedLayer,
-        ...(isSameSelection
-          ? {}
-          : {
-              diffContent: null,
-              diffError: null,
-              diffPreviewMode: "diff",
-              diffPreviewContent: null,
-              diffPreviewLoading: false,
-              diffPreviewError: null,
-            }),
-      };
-    }),
-  selectDiffTab: (path, layer) =>
-    set((s) => {
-      const normalizedLayer = layer ?? null;
-      const isSameSelection =
-        s.diffSelectedFile === path && s.diffSelectedLayer === normalizedLayer;
-      if (isSameSelection) return s;
-      return {
-        diffSelectedFile: path,
-        diffSelectedLayer: normalizedLayer,
-        diffContent: null,
-        diffError: null,
-        diffPreviewMode: "diff",
-        diffPreviewContent: null,
-        diffPreviewLoading: false,
-        diffPreviewError: null,
-      };
-    }),
-  closeDiffTab: (workspaceId, path, layer) =>
-    set((s) => {
-      const normalizedLayer = layer ?? null;
-      const existing = s.diffTabsByWorkspace[workspaceId] ?? [];
-      const idx = existing.findIndex(
-        (t) => t.path === path && t.layer === normalizedLayer,
-      );
-      if (idx < 0) return s;
-      const nextWsTabs = existing.slice(0, idx).concat(existing.slice(idx + 1));
-      const wasActive =
-        s.diffSelectedFile === path && s.diffSelectedLayer === normalizedLayer;
-      const updates: Partial<AppState> = {
-        diffTabsByWorkspace: {
-          ...s.diffTabsByWorkspace,
-          [workspaceId]: nextWsTabs,
-        },
-      };
-      if (wasActive) {
-        // Drop active-diff state. AppLayout will fall back to ChatPanel,
-        // which renders whichever session is in selectedSessionIdByWorkspaceId.
-        updates.diffSelectedFile = null;
-        updates.diffSelectedLayer = null;
-        updates.diffContent = null;
-        updates.diffError = null;
-        updates.diffPreviewMode = "diff";
-        updates.diffPreviewContent = null;
-        updates.diffPreviewLoading = false;
-        updates.diffPreviewError = null;
-      }
-      return updates;
-    }),
-
-  // -- SCM --
-  scmSummary: {},
-  scmDetail: null,
-  scmDetailLoading: false,
-  setScmSummary: (wsId, summary) =>
-    set((s) => ({
-      scmSummary: { ...s.scmSummary, [wsId]: summary },
-    })),
-  setScmDetail: (detail) => set({ scmDetail: detail }),
-  setScmDetailLoading: (loading) => set({ scmDetailLoading: loading }),
-
-  // -- Terminal --
-  terminalTabs: {},
-  activeTerminalTabId: {},
-  terminalPanelVisible: false,
-  workspaceTerminalCommands: {},
-  setTerminalTabs: (wsId, tabs) =>
-    set((s) => ({
-      terminalTabs: { ...s.terminalTabs, [wsId]: tabs },
-    })),
-  addTerminalTab: (wsId, tab) =>
-    set((s) => ({
-      terminalTabs: {
-        ...s.terminalTabs,
-        [wsId]: [...(s.terminalTabs[wsId] || []), tab],
-      },
-      activeTerminalTabId: { ...s.activeTerminalTabId, [wsId]: tab.id },
-      terminalPanelVisible: true,
-    })),
-  removeTerminalTab: (wsId, tabId) =>
-    set((s) => {
-      const tabs = (s.terminalTabs[wsId] || []).filter((t) => t.id !== tabId);
-      const wasActive = s.activeTerminalTabId[wsId] === tabId;
-      // Drop the tab's pane tree and active-pane entry. The terminal panel
-      // cleans up xterm instances by observing the terminalTabs map, and
-      // we never want stale pane state to leak if the tab id is later
-      // reused.
-      const nextTrees = { ...s.terminalPaneTrees };
-      delete nextTrees[tabId];
-      const nextActivePane = { ...s.activeTerminalPaneId };
-      delete nextActivePane[tabId];
-      // When the user closes the last tab in the currently-selected
-      // workspace, collapse the terminal panel — leaving an empty panel
-      // mounted looks broken. If they re-open it later the panel's
-      // tab-load effect will auto-create a fresh tab.
-      const hideBecauseEmpty =
-        tabs.length === 0 && s.selectedWorkspaceId === wsId;
-      return {
-        terminalTabs: { ...s.terminalTabs, [wsId]: tabs },
-        activeTerminalTabId: wasActive
-          ? { ...s.activeTerminalTabId, [wsId]: tabs[0]?.id ?? null }
-          : s.activeTerminalTabId,
-        terminalPaneTrees: nextTrees,
-        activeTerminalPaneId: nextActivePane,
-        terminalPanelVisible: hideBecauseEmpty ? false : s.terminalPanelVisible,
-      };
-    }),
-  setActiveTerminalTab: (wsId, id) =>
-    set((s) => ({
-      activeTerminalTabId: { ...s.activeTerminalTabId, [wsId]: id },
-    })),
-  toggleTerminalPanel: () =>
-    set((s) => ({ terminalPanelVisible: !s.terminalPanelVisible })),
-  setWorkspaceRunningCommand: (wsId, ptyId, command) =>
-    set((s) => {
-      const wsMap = { ...(s.workspaceTerminalCommands[wsId] ?? {}) };
-      wsMap[ptyId] = command;
-      return {
-        workspaceTerminalCommands: {
-          ...s.workspaceTerminalCommands,
-          [wsId]: wsMap,
-        },
-      };
-    }),
-  clearWorkspaceRunningCommand: (wsId, ptyId) =>
-    set((s) => {
-      const wsMap = s.workspaceTerminalCommands[wsId];
-      if (!wsMap || !(ptyId in wsMap)) return {};
-      const next = { ...wsMap };
-      delete next[ptyId];
-      const nextOuter = { ...s.workspaceTerminalCommands };
-      if (Object.keys(next).length === 0) {
-        delete nextOuter[wsId];
-      } else {
-        nextOuter[wsId] = next;
-      }
-      return { workspaceTerminalCommands: nextOuter };
-    }),
-
-  // Pane-tree slice. State is ephemeral: if the app restarts, every tab
-  // comes back as a single-leaf tree. See `terminalPaneTree.ts` for the
-  // pure tree operations these setters wrap.
-  terminalPaneTrees: {},
-  activeTerminalPaneId: {},
-  terminalPaneMaxLeaves: 6,
-  ensurePaneTree: (tabId) => {
-    const existing = get().terminalPaneTrees[tabId];
-    if (existing && existing.kind === "leaf") {
-      const stored = get().activeTerminalPaneId[tabId];
-      if (stored !== existing.id) {
-        set((s) => ({
-          activeTerminalPaneId: {
-            ...s.activeTerminalPaneId,
-            [tabId]: existing.id,
-          },
-        }));
-      }
-      return existing.id;
-    }
-    if (existing && existing.kind === "split") {
-      // Preserve the existing split layout. Use the stored active leaf if
-      // it still identifies a leaf in the tree; otherwise fall back to the
-      // leftmost leaf (and backfill activeTerminalPaneId so future reads
-      // don't hit this branch again).
-      const leaves = allPaneLeafIds(existing);
-      const stored = get().activeTerminalPaneId[tabId];
-      const pick = stored && leaves.includes(stored) ? stored : leaves[0];
-      if (pick !== stored) {
-        set((s) => ({
-          activeTerminalPaneId: {
-            ...s.activeTerminalPaneId,
-            [tabId]: pick,
-          },
-        }));
-      }
-      return pick;
-    }
-    const leaf = makePaneLeaf();
-    set((s) => ({
-      terminalPaneTrees: { ...s.terminalPaneTrees, [tabId]: leaf },
-      activeTerminalPaneId: { ...s.activeTerminalPaneId, [tabId]: leaf.id },
-    }));
-    return leaf.id;
-  },
-  setPaneTree: (tabId, tree) =>
-    set((s) => ({
-      terminalPaneTrees: { ...s.terminalPaneTrees, [tabId]: tree },
-    })),
-  splitPane: (tabId, leafId, direction) => {
-    const state = get();
-    const tree = state.terminalPaneTrees[tabId];
-    if (!tree) return null;
-    const cap = state.terminalPaneMaxLeaves;
-    if (countPaneLeaves(tree) >= cap) return null;
-    const { tree: nextTree, newLeafId } = splitLeafInTree(tree, leafId, direction);
-    if (!newLeafId) return null;
-    set((s) => ({
-      terminalPaneTrees: { ...s.terminalPaneTrees, [tabId]: nextTree },
-      // Focus the freshly created pane.
-      activeTerminalPaneId: { ...s.activeTerminalPaneId, [tabId]: newLeafId },
-    }));
-    return newLeafId;
-  },
-  closePane: (tabId, leafId) => {
-    const state = get();
-    const tree = state.terminalPaneTrees[tabId];
-    if (!tree) return null;
-    const { tree: nextTree, closed, promotedLeafId } = closeLeafInTree(tree, leafId);
-    if (!closed || !promotedLeafId) return null;
-    set((s) => ({
-      terminalPaneTrees: { ...s.terminalPaneTrees, [tabId]: nextTree },
-      activeTerminalPaneId: {
-        ...s.activeTerminalPaneId,
-        [tabId]: promotedLeafId,
-      },
-    }));
-    return promotedLeafId;
-  },
-  setActivePane: (tabId, leafId) =>
-    set((s) => ({
-      activeTerminalPaneId: { ...s.activeTerminalPaneId, [tabId]: leafId },
-    })),
-  setPaneSizes: (tabId, splitId, sizes) =>
-    set((s) => {
-      const tree = s.terminalPaneTrees[tabId];
-      if (!tree) return {};
-      const next = updateSizesInTree(tree, splitId, sizes);
-      if (next === tree) return {};
-      return { terminalPaneTrees: { ...s.terminalPaneTrees, [tabId]: next } };
-    }),
-  setPanePtyId: (tabId, leafId, ptyId) =>
-    set((s) => {
-      const tree = s.terminalPaneTrees[tabId];
-      if (!tree) return {};
-      const rewrite = (n: TerminalPaneNode): TerminalPaneNode => {
-        if (n.kind === "leaf") {
-          return n.id === leafId ? { ...n, ptyId, spawnError: null } : n;
-        }
-        const l = rewrite(n.children[0]);
-        const r = rewrite(n.children[1]);
-        if (l === n.children[0] && r === n.children[1]) return n;
-        return { ...n, children: [l, r] };
-      };
-      const next = rewrite(tree);
-      if (next === tree) return {};
-      return { terminalPaneTrees: { ...s.terminalPaneTrees, [tabId]: next } };
-    }),
-  setPaneSpawnError: (tabId, leafId, error) =>
-    set((s) => {
-      const tree = s.terminalPaneTrees[tabId];
-      if (!tree) return {};
-      const rewrite = (n: TerminalPaneNode): TerminalPaneNode => {
-        if (n.kind === "leaf") {
-          if (n.id !== leafId) return n;
-          // Clear ptyId whenever an error is recorded — use `!= null` so
-          // an empty-string error still invalidates ptyId (truthy check
-          // would miss `""` and leave the UI talking to a dead PTY).
-          return {
-            ...n,
-            spawnError: error,
-            ...(error != null ? { ptyId: undefined } : {}),
-          };
-        }
-        const l = rewrite(n.children[0]);
-        const r = rewrite(n.children[1]);
-        if (l === n.children[0] && r === n.children[1]) return n;
-        return { ...n, children: [l, r] };
-      };
-      const next = rewrite(tree);
-      if (next === tree) return {};
-      return { terminalPaneTrees: { ...s.terminalPaneTrees, [tabId]: next } };
-    }),
-
-  // -- UI --
-  metaKeyHeld: false,
-  setMetaKeyHeld: (held) => set({ metaKeyHeld: held }),
-  sidebarVisible: true,
-  rightSidebarVisible: false,
-  sidebarWidth: 260,
-  rightSidebarWidth: 250,
-  terminalHeight: 300,
-  rightSidebarTab: "changes",
-  sidebarGroupBy: "repo",
-  sidebarRepoFilter: "all",
-  sidebarShowArchived: false,
-  repoCollapsed: {},
-  statusGroupCollapsed: {},
-  fuzzyFinderOpen: false,
-  toggleSidebar: () => set((s) => ({ sidebarVisible: !s.sidebarVisible })),
-  toggleRightSidebar: () =>
-    set((s) => ({ rightSidebarVisible: !s.rightSidebarVisible })),
-  setRightSidebarTab: (tab) => set({ rightSidebarTab: tab }),
-  setSidebarWidth: (w) => set({ sidebarWidth: w }),
-  setRightSidebarWidth: (w) => set({ rightSidebarWidth: w }),
-  setTerminalHeight: (h) => set({ terminalHeight: h }),
-  setSidebarGroupBy: (g) => set({ sidebarGroupBy: g }),
-  setSidebarRepoFilter: (id) => set({ sidebarRepoFilter: id }),
-  setSidebarShowArchived: (show) => set({ sidebarShowArchived: show }),
-  toggleRepoCollapsed: (id) =>
-    set((s) => ({
-      repoCollapsed: {
-        ...s.repoCollapsed,
-        [id]: !s.repoCollapsed[id],
-      },
-    })),
-  toggleStatusGroupCollapsed: (id) =>
-    set((s) => ({
-      statusGroupCollapsed: {
-        ...s.statusGroupCollapsed,
-        [id]: !s.statusGroupCollapsed[id],
-      },
-    })),
-  toggleFuzzyFinder: () =>
-    set((s) => ({ fuzzyFinderOpen: !s.fuzzyFinderOpen })),
-  commandPaletteOpen: false,
-  toggleCommandPalette: () =>
-    set((s) => ({ commandPaletteOpen: !s.commandPaletteOpen })),
-
-  // -- Settings page --
-  settingsOpen: false,
-  settingsSection: null,
-  openSettings: (section = "general") =>
-    set((state) => {
-      const nextSection = section === "plugins" && !state.pluginManagementEnabled
-        ? "experimental"
-        : section;
-      return {
-        settingsOpen: true,
-        settingsSection: nextSection,
-        pluginSettingsIntent: nextSection === "plugins" ? null : state.pluginSettingsIntent,
-        pluginSettingsRepoId: nextSection === "plugins" ? null : state.pluginSettingsRepoId,
-        pluginSettingsTab: nextSection === "plugins" ? "available" : state.pluginSettingsTab,
-      };
-    }),
-  closeSettings: () =>
-    set({
-      settingsOpen: false,
-      settingsSection: null,
-      pluginSettingsIntent: null,
-      pluginSettingsRepoId: null,
-    }),
-  setSettingsSection: (section) =>
-    set((state) => {
-      // Claude Code Plugins requires plugin management to be on; when
-      // disabled, fall through to the experimental pane so the setting
-      // stays reachable. The new "plugins" section (Claudette's own
-      // Lua plugins) is always available.
-      const nextSection =
-        section === "claude-code-plugins" && !state.pluginManagementEnabled
-          ? "experimental"
-          : section;
-      const resetMarketplaceIntent = nextSection === "claude-code-plugins";
-      return {
-        settingsSection: nextSection,
-        pluginSettingsIntent: resetMarketplaceIntent
-          ? null
-          : state.pluginSettingsIntent,
-        pluginSettingsRepoId: resetMarketplaceIntent
-          ? null
-          : state.pluginSettingsRepoId,
-        pluginSettingsTab: resetMarketplaceIntent
-          ? "available"
-          : state.pluginSettingsTab,
-      };
-    }),
-  pluginSettingsTab: "available",
-  pluginSettingsRepoId: null,
-  pluginSettingsIntent: null,
-  pluginRefreshToken: 0,
-  openPluginSettings: (intent = {}) =>
-    set((state) => {
-      if (!state.pluginManagementEnabled) {
-        return {};
-      }
-      const mergedIntent: PluginSettingsIntent = {
-        action: intent.action ?? null,
-        repoId: intent.repoId ?? null,
-        scope: intent.scope ?? "user",
-        source: intent.source ?? null,
-        tab: intent.tab ?? state.pluginSettingsTab,
-        target: intent.target ?? null,
-      };
-      return {
-        settingsOpen: true,
-        settingsSection: "claude-code-plugins",
-        pluginSettingsTab: mergedIntent.tab,
-        pluginSettingsRepoId: mergedIntent.repoId,
-        pluginSettingsIntent: mergedIntent,
-      };
-    }),
-  setPluginSettingsTab: (tab) => set({ pluginSettingsTab: tab }),
-  setPluginSettingsRepoId: (repoId) => set({ pluginSettingsRepoId: repoId }),
-  clearPluginSettingsIntent: () => set({ pluginSettingsIntent: null }),
-  voiceProviderFocus: null,
-  focusVoiceProvider: (providerId) => set({ voiceProviderFocus: providerId }),
-  bumpPluginRefreshToken: () =>
-    set((state) => ({ pluginRefreshToken: state.pluginRefreshToken + 1 })),
-
-  // -- Modals --
-  activeModal: null,
-  modalData: {},
-  openModal: (name, data = {}) => set({ activeModal: name, modalData: data }),
-  closeModal: () => set({ activeModal: null, modalData: {} }),
-
-  // -- Chat input prefill (e.g. after rollback) --
-  chatInputPrefill: null,
-  setChatInputPrefill: (text) => set({ chatInputPrefill: text }),
-  pendingAttachmentsPrefill: null,
-  setPendingAttachmentsPrefill: (atts) =>
-    set({ pendingAttachmentsPrefill: atts }),
-
-  // -- Settings --
-  worktreeBaseDir: "",
-  setWorktreeBaseDir: (dir) => set({ worktreeBaseDir: dir }),
-  defaultBranches: {},
-  setDefaultBranches: (branches) => set({ defaultBranches: branches }),
-  terminalFontSize: 11,
-  setTerminalFontSize: (size) => set({ terminalFontSize: size }),
-  uiFontSize: 13,
-  setUiFontSize: (size) => set({ uiFontSize: size }),
-  fontFamilySans: "",
-  setFontFamilySans: (font) => set({ fontFamilySans: font }),
-  fontFamilyMono: "",
-  setFontFamilyMono: (font) => set({ fontFamilyMono: font }),
-  systemFonts: [],
-  setSystemFonts: (fonts) => set({ systemFonts: fonts }),
-  currentThemeId: DEFAULT_THEME_ID,
-  setCurrentThemeId: (id) => set({ currentThemeId: id }),
-  themeMode: "dark",
-  setThemeMode: (mode) => set({ themeMode: mode }),
-  themeDark: DEFAULT_THEME_ID,
-  setThemeDark: (id) => set({ themeDark: id }),
-  themeLight: DEFAULT_LIGHT_THEME_ID,
-  setThemeLight: (id) => set({ themeLight: id }),
-  lastMessages: {},
-  setLastMessages: (msgs) => set({ lastMessages: msgs }),
-
-  // -- Remote Connections --
-  remoteConnections: [],
-  discoveredServers: [],
-  activeRemoteIds: [],
-  setRemoteConnections: (conns) => set({ remoteConnections: conns }),
-  addRemoteConnection: (conn) =>
-    set((s) => ({ remoteConnections: [...s.remoteConnections, conn] })),
-  removeRemoteConnection: (id) =>
-    set((s) => ({
-      remoteConnections: s.remoteConnections.filter((c) => c.id !== id),
-      activeRemoteIds: s.activeRemoteIds.filter((rid) => rid !== id),
-    })),
-  setDiscoveredServers: (servers) => set({ discoveredServers: servers }),
-  setActiveRemoteIds: (ids) => set({ activeRemoteIds: ids }),
-  addActiveRemoteId: (id) =>
-    set((s) => ({
-      activeRemoteIds: s.activeRemoteIds.includes(id)
-        ? s.activeRemoteIds
-        : [...s.activeRemoteIds, id],
-    })),
-  removeActiveRemoteId: (id) =>
-    set((s) => ({
-      activeRemoteIds: s.activeRemoteIds.filter((rid) => rid !== id),
-    })),
-  mergeRemoteData: (connectionId, data) =>
-    set((s) => {
-      // Tag remote repos and workspaces with the connection ID, then merge.
-      const taggedRepos = data.repositories.map((r) => ({
-        ...r,
-        remote_connection_id: connectionId,
-      }));
-      const taggedWorkspaces = data.workspaces.map((w) => ({
-        ...w,
-        remote_connection_id: connectionId,
-      }));
-      // Merge remote repo default branches so review-workflow prompts and any
-      // other UI keyed off `defaultBranches[repo.id]` work for paired servers.
-      // Prune using the repos *previously* stored for this connection so
-      // entries for repos removed from the latest payload don't linger.
-      const previousRemoteRepoIds = new Set(
-        s.repositories
-          .filter((r) => r.remote_connection_id === connectionId)
-          .map((r) => r.id),
-      );
-      const prunedDefaults = Object.fromEntries(
-        Object.entries(s.defaultBranches).filter(
-          ([id]) => !previousRemoteRepoIds.has(id),
-        ),
-      );
-      return {
-        repositories: [
-          ...s.repositories.filter(
-            (r) => r.remote_connection_id !== connectionId,
-          ),
-          ...taggedRepos,
-        ],
-        workspaces: [
-          ...s.workspaces.filter(
-            (w) => w.remote_connection_id !== connectionId,
-          ),
-          ...taggedWorkspaces,
-        ],
-        defaultBranches: { ...prunedDefaults, ...data.default_branches },
-      };
-    }),
-  clearRemoteData: (connectionId) =>
-    set((s) => {
-      const clearedRepoIds = new Set(
-        s.repositories
-          .filter((r) => r.remote_connection_id === connectionId)
-          .map((r) => r.id),
-      );
-      const prunedDefaults = Object.fromEntries(
-        Object.entries(s.defaultBranches).filter(([id]) => !clearedRepoIds.has(id)),
-      );
-      return {
-        repositories: s.repositories.filter(
-          (r) => r.remote_connection_id !== connectionId,
-        ),
-        workspaces: s.workspaces.filter(
-          (w) => w.remote_connection_id !== connectionId,
-        ),
-        defaultBranches: prunedDefaults,
-      };
-    }),
-
-  // -- Local Server --
-  localServerRunning: false,
-  localServerConnectionString: null,
-  setLocalServerRunning: (running) => set({ localServerRunning: running }),
-  setLocalServerConnectionString: (cs) =>
-    set({ localServerConnectionString: cs }),
-
-  // -- MCP Status --
-  mcpStatus: {},
-  setMcpStatus: (repoId, snapshot) =>
-    set((state) => ({
-      mcpStatus: { ...state.mcpStatus, [repoId]: snapshot },
-    })),
-  clearMcpStatus: (repoId) =>
-    set((state) => {
-      const { [repoId]: _, ...rest } = state.mcpStatus;
-      return { mcpStatus: rest };
-    }),
-
-  // -- Detected Apps --
-  detectedApps: [],
-  setDetectedApps: (apps) => set({ detectedApps: apps }),
-
-  // -- Sidebar Display Preferences --
-  showSidebarRunningCommands: false,
-  setShowSidebarRunningCommands: (v) => set({ showSidebarRunningCommands: v }),
-
-  // -- Experimental --
-  usageInsightsEnabled: false,
-  setUsageInsightsEnabled: (enabled) => set({ usageInsightsEnabled: enabled }),
-  pluginManagementEnabled: false,
-  setPluginManagementEnabled: (enabled) =>
-    set((state) => ({
-      pluginManagementEnabled: enabled,
-      settingsSection:
-        !enabled && state.settingsSection === "claude-code-plugins"
-          ? "experimental"
-          : state.settingsSection,
-      pluginSettingsIntent: enabled ? state.pluginSettingsIntent : null,
-      pluginSettingsRepoId: enabled ? state.pluginSettingsRepoId : null,
-      pluginSettingsTab: enabled ? state.pluginSettingsTab : "available",
-    })),
-  disable1mContext: false,
-  setDisable1mContext: (v) => set({ disable1mContext: v }),
-
-  // -- Claude Code Usage --
-  claudeCodeUsage: null,
-  setClaudeCodeUsage: (usage) => set({ claudeCodeUsage: usage }),
-
-  // -- Metrics --
-  dashboardMetrics: null,
-  analyticsMetrics: null,
-  workspaceMetrics: {},
-  metricsError: null,
-  setDashboardMetrics: (metrics) =>
-    set({ dashboardMetrics: metrics, metricsError: null }),
-  setAnalyticsMetrics: (metrics) =>
-    set({ analyticsMetrics: metrics, metricsError: null }),
-  setWorkspaceMetrics: (metrics) => set({ workspaceMetrics: metrics }),
-  fetchDashboardMetrics: async () => {
-    try {
-      const metrics = await getDashboardMetrics();
-      set({ dashboardMetrics: metrics, metricsError: null });
-    } catch (e) {
-      set({ metricsError: String(e) });
-    }
-  },
-  fetchAnalyticsMetrics: async () => {
-    try {
-      const metrics = await getAnalyticsMetrics();
-      set({ analyticsMetrics: metrics, metricsError: null });
-    } catch (e) {
-      set({ metricsError: String(e) });
-    }
-  },
-  fetchWorkspaceMetricsBatch: async (ids) => {
-    if (ids.length === 0) {
-      set({ workspaceMetrics: {}, metricsError: null });
-      return;
-    }
-    try {
-      const metrics = await getWorkspaceMetricsBatch(ids);
-      set({ workspaceMetrics: metrics, metricsError: null });
-    } catch (e) {
-      set({ metricsError: String(e) });
-    }
-  },
-
-  // -- Updater --
-  updateAvailable: false,
-  updateVersion: null,
-  updateDismissed: false,
-  updateInstallWhenIdle: false,
-  updateDownloading: false,
-  updateProgress: 0,
-  updateChannel: "stable",
-  updateError: null,
-  setUpdateAvailable: (available, version) =>
-    set((state) => ({
-      updateAvailable: available,
-      updateVersion: version,
-      updateDismissed:
-        version === state.updateVersion ? state.updateDismissed : false,
-      updateError: null,
-    })),
-  setUpdateDismissed: (dismissed) => set({ updateDismissed: dismissed }),
-  setUpdateInstallWhenIdle: (enabled) =>
-    set({ updateInstallWhenIdle: enabled }),
-  setUpdateDownloading: (downloading) =>
-    set((state) => ({
-      updateDownloading: downloading,
-      updateError: downloading ? null : state.updateError,
-    })),
-  setUpdateProgress: (progress) => set({ updateProgress: progress }),
-  setUpdateChannel: (channel) =>
-    set({
-      updateChannel: channel,
-      updateAvailable: false,
-      updateVersion: null,
-      updateDismissed: false,
-      updateInstallWhenIdle: false,
-      updateError: null,
-    }),
-  setUpdateError: (error) => set({ updateError: error }),
-
-  // -- App info --
-  appVersion: null,
-  setAppVersion: (version) => set({ appVersion: version }),
-
-  // -- Slash commands --
-  slashCommandsByWorkspace: {},
-  setSlashCommands: (wsId, cmds) =>
-    set((s) => ({
-      slashCommandsByWorkspace: { ...s.slashCommandsByWorkspace, [wsId]: cmds },
-    })),
+  createChatSlice,
+  type ChatSlice,
+  type CompletedTurn,
+  type ToolActivity,
+  type TurnUsage,
+} from "./slices/chatSlice";
+import {
+  createCheckpointsSlice,
+  type CheckpointsSlice,
+} from "./slices/checkpointsSlice";
+import { createDiffSlice, type DiffSlice } from "./slices/diffSlice";
+import {
+  createNotificationsSlice,
+  type NotificationsSlice,
+} from "./slices/notificationsSlice";
+import {
+  createPermissionsSlice,
+  type PermissionLevel,
+  type PermissionsSlice,
+} from "./slices/permissionsSlice";
+import {
+  createRepositoriesSlice,
+  type RepositoriesSlice,
+} from "./slices/repositoriesSlice";
+import { createRemoteSlice, type RemoteSlice } from "./slices/remoteSlice";
+import { createScmSlice, type ScmSlice } from "./slices/scmSlice";
+import {
+  createSettingsSlice,
+  type SettingsSlice,
+} from "./slices/settingsSlice";
+import { createSystemSlice, type SystemSlice } from "./slices/systemSlice";
+import {
+  createTerminalSlice,
+  type TerminalSlice,
+} from "./slices/terminalSlice";
+import { createToolbarSlice, type ToolbarSlice } from "./slices/toolbarSlice";
+import { createUiSlice, type UiSlice } from "./slices/uiSlice";
+import {
+  createWorkspacesSlice,
+  type WorkspacesSlice,
+} from "./slices/workspacesSlice";
+
+export type {
+  AgentQuestion,
+  AgentQuestionItem,
+  ChatSearchState,
+  CompletedTurn,
+  PermissionLevel,
+  PlanApproval,
+  ToolActivity,
+  TurnUsage,
+};
+
+export type AppState = RepositoriesSlice &
+  WorkspacesSlice &
+  ChatSessionsSlice &
+  ChatSlice &
+  AgentInteractionSlice &
+  CheckpointsSlice &
+  NotificationsSlice &
+  ToolbarSlice &
+  PermissionsSlice &
+  DiffSlice &
+  TerminalSlice &
+  ScmSlice &
+  UiSlice &
+  SettingsSlice &
+  RemoteSlice &
+  SystemSlice;
+
+export const useAppStore = create<AppState>()((...a) => ({
+  ...createRepositoriesSlice(...a),
+  ...createWorkspacesSlice(...a),
+  ...createChatSessionsSlice(...a),
+  ...createChatSlice(...a),
+  ...createAgentInteractionSlice(...a),
+  ...createCheckpointsSlice(...a),
+  ...createNotificationsSlice(...a),
+  ...createToolbarSlice(...a),
+  ...createPermissionsSlice(...a),
+  ...createDiffSlice(...a),
+  ...createTerminalSlice(...a),
+  ...createScmSlice(...a),
+  ...createUiSlice(...a),
+  ...createSettingsSlice(...a),
+  ...createRemoteSlice(...a),
+  ...createSystemSlice(...a),
 }));
 
 /**
